pyproject.toml
[project]
name = "barmenui"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.0",
    "langchain>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-openai>=0.3.28",
    "langchain-pinecone>=0.2.11",
    "openai>=1.97.0",
    "pinecone>=7.3.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
